Fix test_does_local_branch_have_commit
