+++ README.md
@@ -3,7 +3,7 @@ TODO
 tox -- test/test_repository.py -s --no-cov
 
 # Testing
-Just a single test file: `tox -e unit -- tests/test_root.py -s`
+Just a single test file: `tox -e unit -- tests/test_single.py -s`
 Lint and type check: `tox -e lint`
 Run all tox environments: `tox`
